fix: soldBy & role replaced by res.locals

+++ src/routes/api/delete_item.ts
@@ -9,16 +9,28 @@ const logger = pino();
 const deleteItem = async (req: Request, res: Response) => {
   const itemID = req.body.itemID;
   try {
-    const rowsDeleted = await Item.destroy({
-      where: { id: itemID },
-    });
-    if (rowsDeleted == 1) {
-      return res.status(204).json({
-        status: "success",
-      });
+    const rowToDelete = await Item.findByPk(itemID);
+    if (rowToDelete?.dataValues.id) {
+      const soldBy = rowToDelete?.dataValues.soldBy;
+      if (soldBy === res.locals.user) {
+        await Item.destroy({
+          where: { id: itemID },
+        });
+        return res.status(204).json({
+          status: "success",
+        });
+      } else {
+        return res.status(400).json({
+          status: "error",
+          error: {
+            message: `You cannot perform this action. Item was not created by ${res.locals.user}`,
+            code: 400,
+          },
+        });
+      }
     } else {
       return res.status(400).json({
-        status: "success",
+        status: "error",
         error: {
           message: "Item does not exist",
           code: 400,
